Arduino Programs - ROS_ArduinoEx&ReadPWM50Hz
The 2 programs I provided here are 2 demo programs that I came across - theoretically, you can connect your Arduino to ROS directly and publish/subscribe to different topics, allegedly. I began working on that BEFORE I got the PCA9685 16 Channel PWM Servo Driver Board in today, but since it required me to install additional things on the RasPi image, I began to install things, however a newer version of cmake was needed, which takes FOREVER to install and deal with, not to mention UNINSTALLING the current version of cmake will BREAK the ROS installation, thus installing the new cmake and over-riding the existing version of cmake via just making the new version of cmake the default in the system default (this was all part of a community post on "how to update cmake" pretty much, w/ a warning from another public form if cmake was REMOVED @ all, it would break ROS :) [happy i read that])... in short I tried to get it working and got stuck on the cmake thing since it just took so long for it to set up packages and sometimes skipping set-up of some items, etc etc... I just waited for the board to come in, since it would effectively solve the same issue I am facing - getting ROS connected to hardware PWMs I can use (aka, follow the tutorial to the T, or, deviate and spend even MORE time developing custom code & stuff that I didnt need to make IF i had the board come in on time (and it did). 

The other script, readPWM50Hz.ino is made for connecting an Arduino uno's "2" pin to it, being able to read the change in state of the rise & fall of a signal, allowing me to effectively read the PWM of the signal into something tangible that can then, later on (if someone develops it in the future) can take & convert to a 500Hz signal, fed to the ESC (speed controller) of the RC car, aka, run the command that the current Arduino program I am using to control the speed of the car, uses, @ a specific PWM value (may need a look-up table or some conversion value or something to convert it into the proper PWM output... relative to the input (divide by 4096?), but this was the starting foundation for that concept. 

While that conversion pgm isnt implemented, this will be the start of it. NOTABLE ISSUE: when it disconnects and reconnects (ex. loose wire), the value changes that it receives, since it reads a change in rise and fall. Because of this, values may vary, and incorrectly so, if such an issue arises. To fix this, it is advised that instead of storing a start time on a state change, that it is instead stored on when a rise occurs, and a fall separately. this way (theoretically) after 1 rise occurs, it will wait for another fall. and if disconnected, then it will simply look for the next fall. despite this ONE value reading being completely extraneous, the subsequent values should be within an acceptable and operational margin, thus throwing a filter to ignore such a large value reading (anything passed 4096 I believe, aka the max value I believe),  should be ignored. 

(I have it all logically planned out in my head, just time constraints...)